+++ src/views/Playground/components/Debug.js
@@ -2,7 +2,19 @@ import React from "react";
 import { makeStyles } from "@material-ui/core/styles";
 import consoleIcon from "../../../assets/imgs/console.svg";
 import PropTypes from "prop-types";
-import { lineOutput } from "./helpers";
+
+export function lineOutput(line, wrkspc, key, classes) {
+    return (
+        <div key={key}>
+            <span className={classes.directory}>
+                {`binari@env/wrkspc${wrkspc}`}
+            </span>{" "}
+            {"> "}
+            {line}
+        </div>
+    );
+}
+
 
 const useStyle = makeStyles({
     log: {
